+++ src/api/controllers/devices/device.patch.validation.js
@@ -4,7 +4,7 @@ module.exports = {
   options: { allowUnknownBody: false },
   body: {
     op: Joi.string().valid(['replace']).required(),
-    path: Joi.string().valid(['/description']).required(),
+    path: Joi.string().valid(['/description', '/tags']).required(),
     value: Joi.any().required(),
   },
 };